HDDS-4768. Refactor the repeated exception conversion method into a common util class (#1864)
